Add less-noisy "Experimental" eye rhymes rule
The current Eye rhymes rule emits a conflict if a single syllable looks
like a rhyme.  This means that "Fool Me Once" matches "Tool" even though
the length makes it unlikely they'd be confused.  The experimental
version requires that at least half of the syllables in the subject name
match an existing handle, e.g. 1/2, 2/3, 2/4, 3/5 syllables.

This change also collapses double consonants in a syllable so that
"This Bean" has two rhymes with "Miss Klean", not one.