Bulk copy schedule slots
Copy Slots button appears at the top level and brings up the copy dialog
with all visible positions shown, but nothing selected.  All slots for
a position may be selected, or individual ones may be.  Additionally,
each position in the main UI has a Bulk Copy button that brings up the
copy dialog with just that position as source options.

Copies can be adjusted by an amount of time (defaulting to "from last
year to this year"), can be copied to a new position, and can have a
few slot attributes like description, max, and info set.