People by position report implementation
Positions are shown with a disclosure triangle, initially collapsed
so that the page is easy to scroll.  Positions can be filtered by type,
e.g. to see just HQ positions.  Positions that not everyone gets show
a table of those people, their status, and whether they're on playa.
All-rangers and all-people positions just show the people who DON'T have
the position so as to avoid printing the same 1400 Rangers 18 times.
Positions also show summary counts.  People can be filtered by status.

This implementation should make it significantly easier to get useful
information.  The main downside is that Ember is S L O W when rendering
thousands of objects, so clicking "Expand all" is a good time to nosh on
a tasty snack from Echelon.