Allow assigned inactive positions for editing
When an inactive position is assigned to a person display it in both the
"Show Positions" and "Edit Positions" in the Person view.  This is to
prevent confusion if a person believes they should have a specific
position assigned to them that has since been deactivated.  This also
allows admins to deselect an inactive position from a person.