Evolve FCC handle rule into Inappropriate Words; add Roma slurs
This fixes the immediate request from
https://github.com/burningmantech/ranger-clubhouse-api/issues/1087
though I imagine we'll want to add other ethnic slurs and words you
shouldn't say on the radio at some future date.  Doing so is now a
simple matter of creating a reason string and associating each
inappropriate word with that reason in the inappropriateWords object.