handle-rules: look for substrings that are inappropriate (#1506)
Without this change, the checker is only looking for exact words
that exist in a proposed handle. Example cases:

Triggered before and after this change:
Shit
Dog Shit

Triggers only after this change, not before:
PissShit
GypsyMan
BassHitter

"BassHitter" is given as an example of one that's probably OK.
All of the inappropriate words are long enough that I wouldn't
expect all that many false positive hits.