handle-reservations: hide expired reservations by default (#1569)
VCs ought to usually just care about active reservations, but it'll be
nice to still be able to edit expired ones, e.g. to extend their end
dates.

This also adds a relevant tweak and TODO to the handle checker tool.
We'll need to deal with priorities a bit better when processing reservations
from the backend.