Add time of day to PNV Alpha message on the dashboard
Noon Pacific is the standard timing for release of Alpha shifts and we seem to get a message from at least one Alpha a week asking why they didn't open at midnight, despite the time being specified in other places. This attempts to stop those by including a time in the PNV message. I'm open to the feedback that this is not the way to do that.