Ensure a browser window showing bookmarks is not affected by login/logout
ember-simple-auth uses events to watch for updated to the local storage. This enables the addon to login and logout other browser windows if need be.

Since we're showing a bookmark, we want to kill the login just for the current browser window, and not login or logout if another window does so.